refactor(Auth): AuthServiceImpl SRP 적용, 포트-어댑터 + 커맨드 패턴 변경
