+++ stempo-application/src/main/java/com/stempo/service/Command.java
@@ -1,0 +1,6 @@
+package com.stempo.service;
+
+public interface Command<T> {
+
+    T execute();
+}
